refactor: class 파일명 수정(클린코드)
